requirements.txt
dotenv==0.9.9
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
